- Added option to return the private key when retrieving certificates.
